+++ app/controllers/solid_stack_web/recurring_tasks_controller.rb
@@ -1,0 +1,7 @@
+module SolidStackWeb
+  class RecurringTasksController < ApplicationController
+    def index
+      @recurring_tasks = SolidQueue::RecurringTask.includes(:recurring_executions).order(:key)
+    end
+  end
+end